Added type prefixes to all @name output.
For example, S_ for Section, LI for ListItems.  THis allows better identification of context, and allows numbers to start the eCC shortName, since there is always a letter prefix.  (@name is intended for use for programming variables, so we have to ensure good @name syntax.  Programming languages do not like variables that start with numbers, or that contain most special characters. )