Correct maxSelections bug in xsd2code serializer
ListFieldType class uses short datatype  for @minSelections and @maxSelections.  The default value is  for both @minSelections and @maxSelections is 1.

However, the shouldSerialize method for both attributes did not use the default values of 1, rather the methods used "default(short)", which is zero.

For  @minSelections , teh minimum value is 1, so the bug never shows up there.  For @maxSelections, teh min value is 0, so we see the bug when @maxSelections =0.

When the value of @maxSelections = 0 for QM, @maxSelections was not serialized into the XML, because the serializer thinks that 0 is the default value of @maxSelections , and default values are set to not be serialized .

The kludge fix was to change the shouldSerializemaxSelections  method to use the correct default values (1).  THsi problem will be regenerated whenever the Schema classes are regenerated by xsd2code.